Further auth API updates/restructure
